refacto: move decorators in utils folder
